requirements.txt
numpy~=2.3.2
llama-cpp-python~=0.3.16
pandas~=2.3.1
